requirements.txt
aiohttp==3.9.1
asyncio-throttle==1.0.2
asyncpraw==7.7.1
beautifulsoup4==4.12.2
feedparser==6.0.10
openai==1.6.1
pydantic==2.5.2
python-dotenv==1.0.0
PyYAML==6.0.1
scikit-learn==1.3.2
slack-sdk==3.26.1
tavily-python==0.3.3
tweepy==4.14.0